changed to size_t2

+++ 0x13-more_singly_linked_lists/1-listint_len.c
@@ -9,7 +9,7 @@
 
 size_t listint_len(const listint_t *h)
 {
-	size_t count;
+	int count;
 
 	while (h)
 	{
